+++ filament/src/fg/fg/RenderTarget.cpp
@@ -16,6 +16,12 @@
 
 #include "RenderTarget.h"
 
+#include "fg/FrameGraph.h"
+
+#include "RenderTargetResource.h"
+
+#include "details/Texture.h"
+
 #include <backend/DriverEnums.h>
 
 namespace filament {
@@ -63,8 +69,13 @@ void RenderTarget::resolve(FrameGraph& fg) noexcept {
                         fg.getResourceEntryUnchecked(attachment.getHandle());
                 // update usage flags for referenced attachments
                 entry.descriptor.usage |= usages[i];
-                // update attachment sample count if not specified
-                entry.descriptor.samples = entry.descriptor.samples ? entry.descriptor.samples : desc.samples;
+
+                // update attachment sample count if not specified and usage permits it
+                if (!entry.descriptor.samples &&
+                    none(entry.descriptor.usage & backend::TextureUsage::SAMPLEABLE)) {
+                    entry.descriptor.samples = desc.samples;
+                }
+
                 attachments |= flags[i];
 
                 // figure out the min/max dimensions across all attachments